extract TIMEUNITS to timeunit.ts
