+++ src/timeunit.ts
@@ -1,0 +1,3 @@
+export const TIMEUNITS = [
+  'year', 'month', 'day', 'date', 'hours', 'minutes', 'seconds'
+];
